Repack Code for MoreOOP
Code and logic was restructured into 4 new classes

DukeMain: Contains the main application and main function to run

DukeParser: Handles logic and understanding of user input

DukeUI: Handles User Interface related functions

DukeStorage: Handles Storage related functions to retrieve and save files
