--- localList.txt
@@ -1,3 +1,0 @@
-T O off myself
-E O off myself but scheduled 2021-08-30
-D O off myself but with a deadline 2012-01-21
